requirements.txt
fastapi
uvicorn
python-dotenv
streamlit
requests
pydantic
yfinance
pandas
reportlab
pillow
